requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
